Refactoring: Move `mod osm` into `pipeline`
No material changes, just code reorganization.
