+++ src/pipeline/osm/cover.rs
